requirements.txt
asgiref==3.9.1
dj-database-url==2.3.0
Django==5.2.6
django-filter==24.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
gunicorn==22.0.0
packaging==25.0
pillow==11.3.0
PyJWT==2.10.1
PyMySQL==1.1.1
pytz==2025.2
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
whitenoise==6.7.0
